\name{mixed-internal}

\alias{names_values}
\alias{sort_labelled}
\alias{order_labelled}
\alias{to_labels}
\alias{order_tagged}
\alias{sort_tagged}
\alias{format_tagged}
\alias{is_tagged_string}

\title{Internal functions}

\description{
  Functions for internal use only.
}

\keyword{internal}
