% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_finder.R
\name{point_finder}
\alias{point_finder}
\title{Point finder}
\usage{
point_finder(drink, area_pcts)
}
\arguments{
\item{drink}{= a dataframe containing two columns, drink and density, maximumn rows  = 4}

\item{area_pct}{a vector of fractional areas, must sum to 1, maximum length = 4}
}
\value{
the points defining the verticies of trapezoids based on drink area and density
}
\description{
Finds points defining area of each liquid
}
\examples{
drinks <- data.frame(drink = c("soda", "cranberry", "milk"),
                     density = c(1, 2, 3),
                     stringsAsFactors = FALSE)

area_pcts <- c(0.25, 0.25, 0.5)
df <- point_finder(drink = drinks, area_pcts = area_pcts)
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
