% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour.R
\name{pour}
\alias{pour}
\title{Pouring a Drink}
\usage{
pour(drink, area_pcts)
}
\arguments{
\item{drink}{= a dataframe containing two columns, drink and density, maximumn rows  = 4}

\item{area_pct}{a vector of fractional areas, must sum to 1, maximum length = 4}
}
\value{
a plot of the layered beverage
}
\description{
Plots layered drinks based on specific gravity
}
\examples{
drinks <- data.frame(drink = c("soda", "cranberry", "milk"),
                     density = c(1, 2, 3),
                     stringsAsFactors = FALSE)
area_pcts <- c(0.25, 0.25, 0.5)
pour(drinks, area_pcts = area_pcts)

}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
