% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_solver.R
\name{quadratic_solver}
\alias{quadratic_solver}
\title{Quadratic Equation Solver}
\usage{
quadratic_solver(a, b, c)
}
\arguments{
\item{a}{the quadratic coefficient}

\item{b}{the linear coefficent}

\item{c}{a constant}
}
\value{
the area of a trapezoid defined by \code{a}, \code{b} and \code{h}
}
\description{
Finds roots of quadratic equations of the form ax^2 + bx + c = 0
}
\examples{
quadratic_solver(1, -3, -10)
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
