% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap_area.R
\name{trap_area}
\alias{trap_area}
\title{Area of a trapezoid}
\usage{
trap_area(a, b, h)
}
\arguments{
\item{a}{the length of one trapezoid face}

\item{b}{the length of the other trapezoid face}

\item{h}{the height of the trapezoid}
}
\value{
the area of a trapezoid defined by \code{a}, \code{b} and \code{h}
}
\description{
Calculates the area of a trapezoid
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
