\name{mixed-internal}

\alias{names_values}
\alias{sort_labelled}
\alias{order_labelled}
\alias{to_labels}
\alias{order_tagged}
\alias{sort_tagged}
\alias{tag_na}
\alias{has_tag}
\alias{get_tag}

\title{Internal functions}

\description{
  Functions for internal use only.
}

\keyword{internal}
