\name{mixed-package}

\alias{mixed-package}

\alias{mixed}

\docType{package}

\title{
Mixed labelled data
}

\description{
A package that extend the \code{haven_labelled} class introduced in package
\pkg{haven}, to deal with mixed labelled objects where existing values that
are defined as missing are treated as both real and missing values, at the
same time.

It also extends the functions related to tagged NA values, also accepting
numbers in addition to single letters.

This package strives for as much compatibility as possible with packages
\pkg{haven} and \pkg{labelled}, adding functionality on top of these two
packages.
}

\details{
    \tabular{ll}{
    Package: \tab mixed\cr
    Type: \tab Package\cr
    Version: \tab 0.7\cr
    Date: \tab 2021-05-13\cr
    License: \tab GPL-v3\cr
    }
}
\author{
Adrian Dusa

Maintainer: Adrian Dusa (dusa.adrian@unibuc.ro)
}

\keyword{misc}
